Feature/lani 791 file import store progress and resume (#153)
* Updates to record last file import file part and records imported.

* Update tests for file import progress

* Added liquibase migration for new cts_file_import columns, rows_imported and last_file_part_imported.

* Format updates

* Minor fix for RowsImported, changed from nullable.

* FileImport endpoint and dbcommand helper updates

* Minor for for null handling in dbhelper fileimport insert method.

* Format update

* Updates for FileImport command handler for RowsImported and LastFilePartImported default values

* Minor code refactor based on SinarQube analysis

* Added missing cancellationtoken parameter to SaveChanges in ProcessFileAsync

* Updated unit tests to consider StartAfterKey parameter with S3 ListObjects method

* LANI-791: Fixed Mark's bad tests

---------

+++ src/Modules/Cads.Cds.StorageBridge/Cads.Cds.StorageBridge.Application/S3Import/Services/IS3ToPostgresService.cs
@@ -5,5 +5,5 @@ namespace Cads.Cds.StorageBridge.Application.S3Import.Services;
 public interface IS3ToPostgresService<T>
     where T : CreateS3ImportJobDto
 {
-    Task<int> ExecuteAsync(T job, CancellationToken cancellationToken = default);
+    Task<long> ExecuteAsync(T job, CancellationToken cancellationToken = default);
 }
